systick: systick_stop bug fix
Test Plan: No

Reviewers: jwzhang, mingzhou

Reviewed By: mingzhou

Differential Revision: http://10.10.20.8/D916

+++ dev/module/ble/ll/evt_ctrl.h
@@ -89,6 +89,6 @@ void ble_timer2_irq2_unmask(void);
 
 void ble_timer2_irq2_mask(void);
 
-void aes_128_start(uint8_t *data,uint8_t *key);
+void aes_128_start(uint8_t *data,uint32_t *key);
 #endif
